Added Utility Class 'Note.cs'
'Note.cs' as the name would imply is a simple script that you can attach to any object (or even as a required component to certain types of scripts) and type in notes of what you plan to do, or remind yourself of what that one complicated script does. It works for any project, honestly.